Handling errors better and not defaulting to open
This changes the default behavior if the required envirionment variables are not set it is locked down.  Defaults are important, this prevents the application from accidentally allowing unauthenticated access.

This also fixes several of the error responses which were uncovered to be returning 500 errors but now return the correct response.