Major overhaul of the test cases
Refactored to have a common set of test cases that test every attribute to make sure both the request and response are valid.

This did require making a defensive copy of the Lock object for the parameter that is passed in because it isn't immutable.  I'd rather this be immutable but this works for now.

These checks the request and response attributes much more throughly.