Using consistgen for the time and UUID dependency injection
This project made me realize that I had implemented the same pattern twice, and as far as I could tell there wasn't an Open Source option or official Java Option that would provide a no-string attached way to have dependency injection for generating timestamps, UUIDs, and random strings.

This pattern was also used in ServiceAuthCentral so two points make a line and extracting this out into a common library is a useful thing.  This is a trivial bit of functionality but involves quite a few classes if you want to have test coverage.